requirements.txt
click==6.7
Flask==1.0.2
Flask-Cors==3.0.6
gunicorn==19.8.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
PyJWT==1.6.4
pymongo==3.6.1
python-dateutil==2.7.3
six==1.11.0
Werkzeug==0.14.1
